+++ src/main/java/com/gscorp/dv1/patrol/web/dto/PatrolCheckpointDto.java
@@ -1,12 +1,13 @@
 package com.gscorp.dv1.patrol.web.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PatrolCheckpointDto (
     UUID externalId,
     String name,
-    Double latitude,
-    Double longitude,
+    BigDecimal latitude,
+    BigDecimal longitude,
     Integer minutesToReach,
     Boolean active
 ){
